requirements.txt
Django
aioredis==2.0.1
python-dotenv==0.21.0
